version 3.02
What's new in release 3.02
Release date: 11 December 2004

  * Significant performance improvements (Jan Jaeger, Greg Smith, Gabor Hoffer)
  * SIE performance almost the same as native (Jan Jaeger, Gabor Hoffer)
  * SCSI tape support in Windows (David "Fish" Trout)
  * Mac OS X CTC networking support (Jay Maynard)
  * Suspend/resume facility (Greg Smith)
  * ASN-and-LX-Reuse Facility (Roger Bowler)
  * Enable or disable ASN-and-LX-reuse in config (Jan Jaeger, Ivan Warren)
  * Extended Translation Facility 3 (Bernard van der Helm)
  * DAT-enhancement facility (Roger Bowler)
  * Immediate CCWs now correctly handled when Suppress Incorrect Length
Indication is specified (Ivan Warren)
  * 3270 option provided to control connection to group of devices (Ivan
Warren)
  * 3270 connections can be limited by IP address (David "Fish" Trout)
  * Remaining 26 binary floating point instructions (Roger Bowler)
  * IPL CLear, System Reset, and System Reset Clear operator commands (Ivan
Warren)
  * Pentium 4 optimizations enabled in gcc (Ivan Warren)

+++ hercifc.h
@@ -2,8 +2,8 @@
 // Hercules Interface Control Program
 // ====================================================================
 //
-// Copyright    (C) Copyright Roger Bowler, 2000-2003
-//              (C) Copyright James A. Pierson, 2002-2003
+// Copyright    (C) Copyright Roger Bowler, 2000-2004
+//              (C) Copyright James A. Pierson, 2002-2004
 //
 
 #ifndef __HERCIFC_H_
@@ -43,14 +43,14 @@
 
 typedef struct _CTLREQ
 {
-    long   iType;
-    int    iProcID;
-    int    iCtlOp;
-    char   szIFName[IFNAMSIZ];
+    long                 iType;
+    int                  iProcID;
+    unsigned long int    iCtlOp;
+    char                 szIFName[IFNAMSIZ];
     union 
     {
-        struct ifreq   ifreq;
-        struct rtentry rtentry;
+        struct ifreq     ifreq;
+        struct rtentry   rtentry;
     } iru;
 }   CTLREQ, *PCTLREQ;
